backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
aioredis==2.0.1
python-dotenv==1.0.0
pydantic==2.5.2
aiohttp==3.8.5